create hash helper module
